Moved utils to root dir -- main.py uses functions from it, so like constants, I think it should be at the same level
