+++ airbrakes/data_handling/data_processor.py
@@ -8,7 +8,7 @@ import numpy.typing as npt
 
 from airbrakes.data_handling.imu_data_packet import EstimatedDataPacket
 from airbrakes.data_handling.processed_data_packet import ProcessedDataPacket
-from airbrakes.utils import deadband
+from utils import deadband
 from constants import ACCELERATION_NOISE_THRESHOLD
