+++ main.py
@@ -12,7 +12,7 @@ from airbrakes.data_handling.logger import Logger
 from airbrakes.hardware.imu import IMU
 from airbrakes.hardware.servo import Servo
 from airbrakes.mock.mock_imu import MockIMU
-from airbrakes.utils import prepare_process_dict, update_display
+from utils import prepare_process_dict, update_display
 from constants import (
     FREQUENCY,
     LOGS_PATH,
