+++ tests/test_utils.py
@@ -1,6 +1,6 @@
 import pytest
 
-from airbrakes.utils import convert_to_float, convert_to_nanoseconds, deadband
+from utils import convert_to_float, convert_to_nanoseconds, deadband
 
 
 @pytest.mark.parametrize(
